New ubuntu-network-setup role!
The main idea is to move to Netplan, or even networkd directly.

So, why still touch this role?

Well, things are different on Ubuntu 18.04 with this new systemd thing...

The new file /lib/modprobe.d/systemd.conf actually disables bond0 and dummy0
"by default", to re-use them, there is a need to create them! Which is cool.

So, in the end of the day, there is no need to use "numdummies" with a fixed
number anymore! No ugly workarounds as well (no /etc/rc.local).

Plus a few clean ups...

In the future, Ansible will manage Netplan and, Netplan will support dummies as
well, then, we can remove ubuntu-network-roles and kiss ifupdown a goodbye.